Refactor postcode_details_client to use a Class
Seperated logic to make a request to the external service from
logic which looks up and returns a part of the search result.
Refactoring as a class also means we’ll no longer need to make
multiple requests if we later want to look up other details of the
postcode returned.